pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spektr"
version = "0.1.0"
description = "Zero-config Python observability. Logging, tracing, and error tracking in one."
requires-python = ">=3.10"
license = "MIT"
readme = "README.md"
authors = [{ name = "Ole Meyer" }]
keywords = ["logging", "tracing", "observability", "opentelemetry", "structured-logging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
]
dependencies = [
    "rich>=13.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.optional-dependencies]
otlp = ["opentelemetry-exporter-otlp-proto-http>=1.20"]
tqdm = ["tqdm>=4.60"]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "starlette>=0.36",
    "fastapi>=0.110",
    "anyio>=4.0",
    "tqdm>=4.60",
]

[project.urls]
Homepage = "https://github.com/olemeyer/spektr"
Repository = "https://github.com/olemeyer/spektr"
Issues = "https://github.com/olemeyer/spektr/issues"
Documentation = "https://github.com/olemeyer/spektr/tree/main/docs"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["no_otel_setup: skip automatic OTel fixture setup"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
]
